requirements.txt
crewai==0.130.0
crewai-tools==0.47.1
python-dotenv
uvicorn==0.29.0
fastapi==0.110.3
pydantic==2.7.1
pydantic-core==2.18.2
aiofiles
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.9
langchain-core==0.1.52
langsmith==0.1.67
onnxruntime==1.22.0
openai==1.30.5
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0